refactor: move middleRowFocus logic to module
